# openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestionOut"
            },
            "title": "Questions",
            "type": "array"
          },
          "results": {
            "additionalProperties": true,
            "title": "Results",
            "type": "object"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "title",
          "questions",
          "score",
          "results"
        ],
        "title": "AssessmentOut",
        "type": "object"
      },
      "AssessmentQuestionOut": {
        "properties": {
          "concept": {
            "title": "Concept",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "index",
          "type",
          "prompt",
          "concept"
        ],
        "title": "AssessmentQuestionOut",
        "type": "object"
      },
      "AssessmentSubmitIn": {
        "properties": {
          "responses": {
            "items": {},
            "maxItems": 12,
            "title": "Responses",
            "type": "array"
          }
        },
        "required": [
          "responses"
        ],
        "title": "AssessmentSubmitIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JourneyConceptOut": {
        "properties": {
          "evidence": {
            "title": "Evidence",
            "type": "integer"
          },
          "misconceptions": {
            "items": {
              "type": "string"
            },
            "title": "Misconceptions",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "name",
          "state",
          "evidence",
          "misconceptions"
        ],
        "title": "JourneyConceptOut",
        "type": "object"
      },
      "JourneyLessonOut": {
        "properties": {
          "concepts": {
            "items": {
              "type": "string"
            },
            "title": "Concepts",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "concepts"
        ],
        "title": "JourneyLessonOut",
        "type": "object"
      },
      "JourneyModuleOut": {
        "properties": {
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/JourneyLessonOut"
            },
            "title": "Lessons",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "lessons"
        ],
        "title": "JourneyModuleOut",
        "type": "object"
      },
      "JourneyOut": {
        "properties": {
          "checkpoints": {
            "title": "Checkpoints",
            "type": "integer"
          },
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/JourneyConceptOut"
            },
            "title": "Concepts",
            "type": "array"
          },
          "current": {
            "$ref": "#/components/schemas/JourneyPositionOut"
          },
          "focus_level": {
            "title": "Focus Level",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "memory": {
            "items": {
              "$ref": "#/components/schemas/MemorySummaryOut"
            },
            "title": "Memory",
            "type": "array"
          },
          "momentum": {
            "$ref": "#/components/schemas/MomentumOut"
          },
          "objective": {
            "title": "Objective",
            "type": "string"
          },
          "parked": {
            "items": {
              "type": "string"
            },
            "title": "Parked",
            "type": "array"
          },
          "plan": {
            "items": {
              "$ref": "#/components/schemas/JourneyModuleOut"
            },
            "title": "Plan",
            "type": "array"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subject",
          "objective",
          "level",
          "status",
          "plan",
          "current",
          "checkpoints",
          "concepts",
          "session_id",
          "memory",
          "parked",
          "focus_level",
          "momentum"
        ],
        "title": "JourneyOut",
        "type": "object"
      },
      "JourneyPositionOut": {
        "properties": {
          "concept": {
            "title": "Concept",
            "type": "string"
          },
          "lesson": {
            "title": "Lesson",
            "type": "integer"
          },
          "module": {
            "title": "Module",
            "type": "integer"
          }
        },
        "required": [
          "module",
          "lesson",
          "concept"
        ],
        "title": "JourneyPositionOut",
        "type": "object"
      },
...
      "MemorySummaryOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "turn_from": {
            "title": "Turn From",
            "type": "integer"
          },
          "turn_to": {
            "title": "Turn To",
            "type": "integer"
          }
        },
        "required": [
          "level",
          "turn_from",
          "turn_to",
          "summary",
          "created_at"
        ],
        "title": "MemorySummaryOut",
        "type": "object"
      },
...
      "MomentumOut": {
        "description": "Today, plainly: showings recorded and concepts that reached mastered.\nA count, not a score — no streaks, no anxiety.",
        "properties": {
          "events_today": {
            "title": "Events Today",
            "type": "integer"
          },
          "mastered_today": {
            "title": "Mastered Today",
            "type": "integer"
          }
        },
        "required": [
          "events_today",
          "mastered_today"
        ],
        "title": "MomentumOut",
        "type": "object"
      },
...
      "RecallIn": {
        "properties": {
          "elapsed_ms": {
            "default": 0,
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "type": "integer"
          },
          "rating": {
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "rating"
        ],
        "title": "RecallIn",
        "type": "object"
      },
      "RecallOut": {
        "properties": {
          "card_id": {
            "format": "uuid",
            "title": "Card Id",
            "type": "string"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "scheduled_days": {
            "title": "Scheduled Days",
            "type": "number"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          }
        },
        "required": [
          "card_id",
          "due_at",
          "scheduled_days",
          "concept",
          "state"
        ],
        "title": "RecallOut",
        "type": "object"
      },
      "RecapOut": {
        "description": "The ten-second recap (\"resume pra mim\"), from state alone.",
        "properties": {
          "know": {
            "items": {
              "type": "string"
            },
            "title": "Know",
            "type": "array"
          },
          "next": {
            "items": {
              "type": "string"
            },
            "title": "Next",
            "type": "array"
          },
          "now": {
            "title": "Now",
            "type": "string"
          },
          "parked": {
            "items": {
              "type": "string"
            },
            "title": "Parked",
            "type": "array"
          },
          "shaky": {
            "items": {
              "type": "string"
            },
            "title": "Shaky",
            "type": "array"
          }
        },
        "required": [
          "know",
          "shaky",
          "now",
          "next",
          "parked"
        ],
        "title": "RecapOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "NOEMA — open-source adaptive learning platform.\n\nCookie sessions authenticate the web app; mutations require the CSRF token returned\nat login in the `x-csrf-token` header.",
    "title": "NOEMA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/assessments/{assessment_id}": {
      "get": {
        "operationId": "get_assessment_api_v1_ai_assessments__assessment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assessment",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/assessments/{assessment_id}/submit": {
      "post": {
        "description": "Hand the paper in. Closed questions are graded here and now; open ones\nby the rubric grader on the economy tier. The results feed the student\nmodel; the correction is the next Professor turn.",
        "operationId": "submit_assessment_api_v1_ai_assessments__assessment_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmitIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Assessment",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/journeys/latest": {
      "get": {
        "description": "The journey the learner was last on — the home screen's continue.",
        "operationId": "latest_journey_api_v1_ai_journeys_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "notebook_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notebook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JourneyOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Journey Api V1 Ai Journeys Latest Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Journey",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/journeys/{journey_id}": {
      "get": {
        "operationId": "get_journey_api_v1_ai_journeys__journey_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Journey Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Journey",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/journeys/{journey_id}/cards/{card_id}/recall": {
      "post": {
        "description": "A lesson card turned over and graded — approves it on first reading,\nrecords the FSRS review and a mastery event.",
        "operationId": "recall_card_api_v1_ai_journeys__journey_id__cards__card_id__recall_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "card_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Card Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recall Card",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/journeys/{journey_id}/recap": {
      "get": {
        "description": "YOU KNOW · NOW · NEXT, in ten seconds, without a model call.",
        "operationId": "journey_recap_api_v1_ai_journeys__journey_id__recap_get",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Journey Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecapOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journey Recap",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}